Plug : Move iterators from PlugIterator.h to Plug.h.
All our other GraphComponent types declare their iterators in the same header - doing the same for plugs improves consistency.
