+++ include/Gaffer/Metadata.h
@@ -38,6 +38,7 @@
 #define GAFFER_METADATA_H
 
 #include "boost/function.hpp"
+#include "boost/signals.hpp"
 
 #include "IECore/InternedString.h"
 #include "IECore/Data.h"
